+++ HttpRetriesApp/HttpRetriesApp.csproj
@@ -12,6 +12,7 @@
     </ItemGroup>
 
     <ItemGroup>
+      <PackageReference Include="BenchmarkDotNet" Version="0.13.1" />
       <PackageReference Include="Polly.Extensions.Http" Version="3.0.0" />
     </ItemGroup>
